+++ include/engine/trip/trip_brute_force.hpp
@@ -31,7 +31,20 @@ EdgeWeight ReturnDistance(const util::DistTableWrapper<EdgeWeight> &dist_table,
     std::size_t i = 0;
     while (i < location_order.size() && (route_dist < min_route_dist))
     {
-        route_dist += dist_table(location_order[i], location_order[(i + 1) % component_size]);
+        // Check for path that touches INVALID_EDGE_WEIGHT (which is MAX_32INT)
+        // Fixed start and end trip needs to escape this path if a max edge weight is found on the
+        // dist table
+        if (dist_table(location_order[i], location_order[(i + 1) % component_size]) ==
+            INVALID_EDGE_WEIGHT)
+        {
+            return INVALID_EDGE_WEIGHT;
+        }
+        else
+        {
+            route_dist += dist_table(location_order[i], location_order[(i + 1) % component_size]);
+        }
+
+        // This boost assert should not be reached if TFSE table
         BOOST_ASSERT_MSG(dist_table(location_order[i], location_order[(i + 1) % component_size]) !=
                              INVALID_EDGE_WEIGHT,
                          "invalid route found");